% Auto-generated: do not edit by hand
\name{markdownInputComponent}

\alias{markdownInputComponent}

\title{MarkdownInputComponent component}

\description{
MarkdownInputComponent is a wrapper on the SimpleMDE component. It renders an input with the property `value` which is editable by the user.
}

\usage{
markdownInputComponent(id=NULL, value=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{value}{Character. The value displayed in the input.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

