% Auto-generated: do not edit by hand
\name{markdownInputComponent}

\alias{markdownInputComponent}

\title{MarkdownInputComponent component}

\description{
MarkdownInputComponent for Dash. Provides a rich text editor for Markdown input with live preview.
}

\usage{
markdownInputComponent(id=NULL, className=NULL, options=NULL, style=NULL,
value=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{className}{Character. Additional CSS class for the container.}

\item{options}{Named list. Options for the SimpleMDE editor.
See: https://github.com/sparksuite/simplemde-markdown-editor#configuration}

\item{style}{Named list. Inline style for the container.}

\item{value}{Character. The Markdown content.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

